backend/requirements.txt
django
django-mptt
pyjwt